+++ internal/repository/memory/MemStorage_test.go
@@ -131,7 +131,7 @@ func TestMemStorage_Get(t *testing.T) {
 func TestMemStorage_GetAll(t *testing.T) {
 	metrics := models.Metrics{}
 	metrics.Update("Alloc", models.GaugeType, 101.42)
-	metrics.Update("PollCount", models.CounterType, int64(2))
+	metrics.Update("PollCount", models.CounterType, 2)
 
 	type fields struct {
 		metrics models.Metrics
